+++ tests/src/integration/derpibooru-test.cpp
@@ -9,7 +9,7 @@ void DerpibooruTest::testHtml()
 	QList<Image*> images = getImages("Booru-on-rails", "derpibooru.org", "regex", "safe", "results.html");
 
 	// Convert results
-	QList<int> ids;
+	QList<qulonglong> ids;
 	for (Image *img : images)
 	{
 		ids.append(img->id());
@@ -17,7 +17,7 @@ void DerpibooruTest::testHtml()
 
 	// Check results
 	ids = ids.mid(0, 3);
-	QList<int> expected = QList<int>() << 1334191 << 1334187 << 1334186;
+	QList<qulonglong> expected = QList<qulonglong>() << 1334191 << 1334187 << 1334186;
 	QCOMPARE(images.count(), 15);
 	QCOMPARE(ids, expected);
 }
@@ -27,7 +27,7 @@ void DerpibooruTest::testJson()
 	QList<Image*> images = getImages("Booru-on-rails", "derpibooru.org", "json", "safe", "results.json");
 
 	// Convert results
-	QList<int> ids;
+	QList<qulonglong> ids;
 	for (Image *img : images)
 	{
 		ids.append(img->id());
@@ -35,7 +35,7 @@ void DerpibooruTest::testJson()
 
 	// Check results
 	ids = ids.mid(0, 3);
-	QList<int> expected = QList<int>() << 1248664 << 1248663 << 1248661;
+	QList<qulonglong> expected = QList<qulonglong>() << 1248664 << 1248663 << 1248661;
 	QCOMPARE(images.count(), 15);
 	QCOMPARE(ids, expected);
 }
